requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
python-multipart
requests
pydantic
tqdm
python-dotenv
gunicorn
groq